Refactored the project files into respective and a more sensible folder-structure
