+++ src/properties.py
@@ -1,5 +1,5 @@
 # Non-MacOS users can change it to Chrome/Firefox.
-BROWSER = "Safari" # can be Chrome/Safari/Firefox
+BROWSER = "Chrome" # can be Chrome/Safari/Firefox
 MATCH_URL = "http://www.espncricinfo.com/series/8039/commentary/1144506/afghanistan-vs-england-24th-match-icc-cricket-world-cup-2019"
 MESSAGE_BOX_CLASS_NAME = "_3u328"
 SEND_BUTTON_CLASS_NAME = "_3M-N-"
