+++ src/utils/TimeUtils.py
@@ -1,0 +1,5 @@
+import datetime
+
+def get_current_time():
+    return datetime.datetime.now()
+
